+++ src/ECU/tests/test_engineHeater.cpp
@@ -200,7 +200,7 @@ void test_heater_state_disables_when_conditions_fail(void) {
 
 int main(void) {
   initSensors();
-  hal_eeprom_init(HAL_EEPROM_RP2040, 512, 0);
+  hal_eeprom_init(HAL_EEPROM_FLASH, 512, 0);
 
   UNITY_BEGIN();